requirements.txt
# 永恒之塔2 守护星辅助脚本依赖
# 模块化版本

pyyaml>=6.0
keyboard>=0.13.5
mouse>=0.7.1
mss>=9.0.1
pyserial>=3.5
numpy>=1.24.0
opencv-python>=4.8.0
